- added give_app_context to init - fixed modular imports
