+++ backend/AudioProcessor.py
@@ -1,8 +1,8 @@
-from backend.word_removal import WordRemover
-from backend.STT import SpeachToText
+from .word_removal import WordRemover
+from .STT import SpeachToText
 from pydub import AudioSegment, effects, silence
 
-from backend.AudioFile import Audio
+from .AudioFile import Audio
 
 class AudioProcessingService:
     def __init__(self, audioFile):
